requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.7
bcrypt>=4.0.1
langgraph>=0.1.0
langchain_community>=0.1.0
pydantic_v1>=2.5.3
python-jose[cryptography]
passlib[bcrypt]
python-multipart
jwt
pydantic[email]
python-multipart
python-jose[cryptography]
passlib[bcrypt]